requirements.txt
# python3.11
camelot-py[cv]
PyPDF2~=2.0
sentence-transformers
streamlit
--upgrade camelot